:sparkles: fix bytes to int32_t. and fix data type for scanning mode.

+++ src/grib_coder/grib_section_3.h
@@ -38,6 +38,6 @@ private:
 	int32_t longitude_of_last_grid_point_;
 	uint32_t i_direction_increment_;
 	uint32_t j_direction_increment_;
-	uint32_t scanning_mode_;
+	uint8_t scanning_mode_;
 };
 } // namespace GribCoder
